+++ clang-tools-extra/test/clang-tidy/checkers/Inputs/readability-identifier-naming/global-style1/header.h
@@ -1,0 +1,5 @@
+
+
+void style_first_good();
+
+void styleFirstBad();